Added direction button logic
Direction buttons will turn off if there is no set location in that direction.  Button will turn back on if there is, and will also set the location "behind" to current location.
ie - Location B is north of location A, which means that location A is south of location B.